+++ src/app/common/utils/counter.ts
@@ -7,6 +7,9 @@ export function createCounter(startPosition = 0) {
     increment() {
       return (count += 1);
     },
+    incrementBy(amount: number) {
+      return (count += amount);
+    },
     decrement() {
       return (count -= 1);
     },