+++ src/app/debug.ts
@@ -1,6 +1,9 @@
 import toast from 'react-hot-toast';
 
+import * as reduxPersist from 'redux-persist';
+
 import { getLogsFromBrowserStorage } from '@shared/logger-storage';
+import { persistConfig } from '@shared/storage';
 
 import { store } from './store';
 import { stxChainSlice } from './store/chains/stx-chain.slice';
@@ -19,6 +22,10 @@ const debug = {
   },
   logStore() {
     return store.getState();
+  },
+  // Utilised in integration tests
+  async logPersistedStore() {
+    return reduxPersist.getStoredState(persistConfig);
   },
   setHighestAccountIndex(index: number) {
     toast.success('Highest account index set to ' + index);