+++ src/app/pages/home/components/account-area.tsx
@@ -4,14 +4,14 @@ import { Stack, StackProps } from '@stacks/ui';
 
 import { CurrentAccountAvatar } from '@app/features/current-account/current-account-avatar';
 import { CurrentAccountName } from '@app/features/current-account/current-account-name';
-import { useCurrentBtcNativeSegwitAccountAddressIndexZero } from '@app/store/accounts/blockchain/bitcoin/native-segwit-account.hooks';
+import { useCurrentAccountNativeSegwitAddressIndexZero } from '@app/store/accounts/blockchain/bitcoin/native-segwit-account.hooks';
 import { useCurrentStacksAccount } from '@app/store/accounts/blockchain/stacks/stacks-account.hooks';
 
 import { AccountTotalBalance } from '../../../components/account-total-balance';
 
 export const CurrentAccount = memo((props: StackProps) => {
   const currentAccount = useCurrentStacksAccount();
-  const btcAddress = useCurrentBtcNativeSegwitAccountAddressIndexZero();
+  const btcAddress = useCurrentAccountNativeSegwitAddressIndexZero();
 
   if (!currentAccount) return null;
   return (