+++ src/app/pages/receive-tokens/receive-btc.tsx
@@ -6,7 +6,7 @@ import get from 'lodash.get';
 
 import { useAnalytics } from '@app/common/hooks/analytics/use-analytics';
 import { useCurrentAccountIndex } from '@app/store/accounts/account';
-import { useBtcNativeSegwitAccountIndexAddressIndexZero } from '@app/store/accounts/blockchain/bitcoin/native-segwit-account.hooks';
+import { useNativeSegwitAccountIndexAddressIndexZero } from '@app/store/accounts/blockchain/bitcoin/native-segwit-account.hooks';
 
 import { ReceiveBtcModalWarning } from './components/receive-btc-warning';
 import { ReceiveTokensLayout } from './components/receive-tokens.layout';
@@ -18,7 +18,7 @@ export function ReceiveBtcModal() {
   const currentAccountIndex = useCurrentAccountIndex();
   const accountIndex = get(state, 'accountIndex', currentAccountIndex);
 
-  const activeAccountBtcAddress = useBtcNativeSegwitAccountIndexAddressIndexZero(accountIndex);
+  const activeAccountBtcAddress = useNativeSegwitAccountIndexAddressIndexZero(accountIndex);
   const btcAddress = get(state, 'btcAddress', activeAccountBtcAddress);
 
   const { onCopy } = useClipboard(btcAddress);