+++ src/app/query/stacks/balance/stacks-ft-balances.hooks.ts
@@ -27,6 +27,9 @@ export function useStacksAnchoredCryptoCurrencyAssetBalance(address: string) {
       createStacksCryptoCurrencyAssetTypeWrapper(parseBalanceResponse(resp).stx.unlockedStx.amount),
   });
 }
+
+// we will probably need this in the future
+// ts-unused-exports:disable-next-line
 export function useStacksUnanchoredCryptoCurrencyAssetBalance(address: string) {
   return useUnanchoredStacksAccountBalanceQuery(address, {
     select: resp =>