+++ src/app/store/accounts/blockchain/bitcoin/native-segwit-account.hooks.ts
@@ -1,23 +1,9 @@
-import { useCallback, useMemo } from 'react';
 import { useSelector } from 'react-redux';
 
-import { HDKey } from '@scure/bip32';
-
-import {
-  deriveAddressIndexKeychainFromAccount,
-  deriveAddressIndexZeroFromAccount,
-} from '@shared/crypto/bitcoin/bitcoin.utils';
-import {
-  deriveNativeSegWitReceiveAddressIndex,
-  getNativeSegWitPaymentFromKeychain,
-} from '@shared/crypto/bitcoin/p2wpkh-address-gen';
-import { isUndefined } from '@shared/utils';
+import { deriveAddressIndexKeychainFromAccount } from '@shared/crypto/bitcoin/bitcoin.utils';
+import { getNativeSegWitPaymentFromAddressIndex } from '@shared/crypto/bitcoin/p2wpkh-address-gen';
 
 import { bitcoinNetworkModeToCoreNetworkMode, whenNetwork } from '@app/common/utils';
-import {
-  formatBitcoinAccount,
-  tempHardwareAccountForTesting,
-} from '@app/store/accounts/blockchain/bitcoin/bitcoin-account.models';
 import { useCurrentNetwork } from '@app/store/networks/networks.selectors';
 
 import { useCurrentAccountIndex } from '../../account';
@@ -25,9 +11,10 @@ import {
   bitcoinSignerFactory,
   selectMainnetNativeSegWitKeychain,
   selectTestnetNativeSegWitKeychain,
+  useMakeBitcoinNetworkSignersForPaymentType,
 } from './bitcoin-keychain';
 
-function useNativeSegwitAccountKeychain() {
+function useNativeSegwitActiveNetworkAccountPrivateKeychain() {
   const network = useCurrentNetwork();
   return useSelector(
     whenNetwork(bitcoinNetworkModeToCoreNetworkMode(network.chain.bitcoin.network))({
@@ -38,111 +25,66 @@ function useNativeSegwitAccountKeychain() {
 }
 
 export function useNativeSegwitCurrentAccountPrivateKeychain() {
-  const keychain = useNativeSegwitAccountKeychain();
+  const keychain = useNativeSegwitActiveNetworkAccountPrivateKeychain();
   const currentAccountIndex = useCurrentAccountIndex();
   return keychain?.(currentAccountIndex);
 }
 
-function useCurrentBitcoinNativeSegwitAccountPublicKeychain() {
-  const { xpub } = useCurrentBitcoinNativeSegwitAccountInfo();
-  if (!xpub) return; // TODO: Revisit this return early
-  const keychain = HDKey.fromExtendedKey(xpub);
-  if (!keychain.publicKey) throw new Error('No public key for given keychain');
-  if (!keychain.pubKeyHash) throw new Error('No pub key hash for given keychain');
-  return keychain;
-}
+export function useNativeSegwitNetworkSigners() {
+  const mainnetKeychainFn = useSelector(selectMainnetNativeSegWitKeychain);
+  const testnetKeychainFn = useSelector(selectTestnetNativeSegWitKeychain);
 
-// Concept of current address index won't exist with privacy mode
-export function useCurrentBitcoinNativeSegwitAddressIndexPublicKeychain() {
-  const keychain = useCurrentBitcoinNativeSegwitAccountPublicKeychain();
-  if (!keychain) return; // TODO: Revisit this return early
-  return deriveAddressIndexZeroFromAccount(keychain);
-}
-
-export function useAllBitcoinNativeSegWitNetworksByAccount() {
-  const mainnetKeychainAtAccount = useSelector(selectMainnetNativeSegWitKeychain);
-  const testnetKeychainAtAccount = useSelector(selectTestnetNativeSegWitKeychain);
-
-  return useCallback(
-    (accountIndex: number) => {
-      if (!mainnetKeychainAtAccount || !testnetKeychainAtAccount)
-        throw new Error('Cannot derive addresses in non-software mode');
-      return {
-        mainnet: deriveNativeSegWitReceiveAddressIndex({
-          xpub: mainnetKeychainAtAccount(accountIndex).publicExtendedKey,
-          network: 'mainnet',
-        })?.address,
-        testnet: deriveNativeSegWitReceiveAddressIndex({
-          xpub: testnetKeychainAtAccount(accountIndex).publicExtendedKey,
-          network: 'testnet',
-        })?.address,
-        regtest: deriveNativeSegWitReceiveAddressIndex({
-          xpub: testnetKeychainAtAccount(accountIndex).publicExtendedKey,
-          network: 'regtest',
-        })?.address,
-      };
-    },
-    [mainnetKeychainAtAccount, testnetKeychainAtAccount]
+  return useMakeBitcoinNetworkSignersForPaymentType(
+    mainnetKeychainFn,
+    testnetKeychainFn,
+    getNativeSegWitPaymentFromAddressIndex
   );
 }
 
-function useBitcoinNativeSegwitAccountInfo(index: number) {
-  const keychain = useNativeSegwitAccountKeychain();
-  return useMemo(() => {
-    // TODO: Remove with bitcoin Ledger integration
-    if (isUndefined(keychain)) return tempHardwareAccountForTesting;
-    return formatBitcoinAccount(keychain(index))(index);
-  }, [keychain, index]);
+function useCurrentAccountNativeSegwitSignerIndexZero() {
+  return useCurrentAccountNativeSegwitSigner()?.(0);
 }
 
-function useCurrentBitcoinNativeSegwitAccountInfo() {
-  const currentAccountIndex = useCurrentAccountIndex();
-  return useBitcoinNativeSegwitAccountInfo(currentAccountIndex);
-}
-
-function useDeriveNativeSegWitAccountIndexAddressIndexZero(xpub: string) {
+function useNativeSegwitSigner(accountIndex: number) {
   const network = useCurrentNetwork();
-  return useMemo(
-    () =>
-      deriveNativeSegWitReceiveAddressIndex({
-        xpub,
-        network: network.chain.bitcoin.network,
-      }),
-    [xpub, network]
-  );
-}
-
-export function useCurrentBtcNativeSegwitAccountAddressIndexZero() {
-  const { xpub } = useCurrentBitcoinNativeSegwitAccountInfo();
-  return useDeriveNativeSegWitAccountIndexAddressIndexZero(xpub)?.address as string;
-}
-
-export function useBtcNativeSegwitAccountIndexAddressIndexZero(accountIndex: number) {
-  const { xpub } = useBitcoinNativeSegwitAccountInfo(accountIndex);
-  return useDeriveNativeSegWitAccountIndexAddressIndexZero(xpub)?.address as string;
-}
-
-export function useCurrentAccountNativeSegwitSigner() {
-  const network = useCurrentNetwork();
-  const index = useCurrentAccountIndex();
-  const accountKeychain = useNativeSegwitAccountKeychain()?.(index);
+  const accountKeychain = useNativeSegwitActiveNetworkAccountPrivateKeychain()?.(accountIndex);
   if (!accountKeychain) return;
-  const addressIndexKeychainFn = deriveAddressIndexKeychainFromAccount(accountKeychain);
 
   return bitcoinSignerFactory({
-    addressIndexKeychainFn,
-    paymentFn: getNativeSegWitPaymentFromKeychain,
+    accountKeychain,
+    paymentFn: getNativeSegWitPaymentFromAddressIndex,
     network: network.chain.bitcoin.network,
   });
 }
 
 export function useCurrentAccountNativeSegwitDetails() {
-  const network = useCurrentNetwork();
-  const index = useCurrentAccountIndex();
-  const accountKeychain = useNativeSegwitAccountKeychain()?.(index);
-  if (!accountKeychain) return;
-  const addressIndexKeychainFn = deriveAddressIndexKeychainFromAccount(accountKeychain);
-  const addressIndexKeychain = addressIndexKeychainFn(index);
+  const currentAccountIndex = useCurrentAccountIndex();
+  const currentNetwork = useCurrentNetwork();
+  const currentAddress = useNativeSegwitAccountIndexAddressIndexZero(currentAccountIndex);
+  const currentAccountKeychain = useNativeSegwitCurrentAccountPrivateKeychain();
+  if (!currentAccountKeychain) return;
+  const currentAddressIndexKeychain =
+    deriveAddressIndexKeychainFromAccount(currentAccountKeychain)(currentAccountIndex);
 
-  return {network, addressIndexKeychain};
+  return { currentNetwork, currentAddress, currentAddressIndexKeychain };
+}
+
+export function useCurrentAccountNativeSegwitSigner() {
+  const currentAccountIndex = useCurrentAccountIndex();
+  return useNativeSegwitSigner(currentAccountIndex);
+}
+
+export function useCurrentAccountNativeSegwitAddressIndexZero() {
+  const signer = useCurrentAccountNativeSegwitSignerIndexZero();
+  return signer?.payment.address as string;
+}
+
+export function useNativeSegwitAccountIndexAddressIndexZero(accountIndex: number) {
+  const signer = useNativeSegwitSigner(accountIndex)?.(0);
+  return signer?.payment.address as string;
+}
+
+export function useCurrentBitcoinNativeSegwitAddressIndexPublicKeychain() {
+  const signer = useCurrentAccountNativeSegwitSignerIndexZero();
+  return signer?.publicKeychain;
 }